Advanced the work on possible moves calculator;
Make it possible to calculate, based on the current values set, all of the possible outcomes and their results.
Performance is affected when choosing to go too deep into this process of calculating ourcomes for a certain possible move, this is partily due to the large number of results, but the impact of the data structure and the functions used also makes a significant impact. Possible improvement of performance here, though not really needed for the moment.